Move matchesToResolvers to the only place it's used
matchesToResolvers is now only used to convert
SearchResultResolver.Results into resolvers, so this commit moves it to
be colocated in that file and unexported.
