+++ cmd/frontend/graphqlbackend/search_results.go
@@ -131,7 +131,44 @@ func (sr *SearchResultsResolver) Results() []SearchResultResolver {
 		limited = sr.SearchResults[:sr.limit]
 	}
 
-	return MatchesToResolvers(sr.db, limited)
+	return matchesToResolvers(sr.db, limited)
+}
+
+func matchesToResolvers(db dbutil.DB, matches []result.Match) []SearchResultResolver {
+	type repoKey struct {
+		Name types.RepoName
+		Rev  string
+	}
+	repoResolvers := make(map[repoKey]*RepositoryResolver, 10)
+	getRepoResolver := func(repoName types.RepoName, rev string) *RepositoryResolver {
+		if existing, ok := repoResolvers[repoKey{repoName, rev}]; ok {
+			return existing
+		}
+		resolver := NewRepositoryResolver(db, repoName.ToRepo())
+		resolver.RepoMatch.Rev = rev
+		repoResolvers[repoKey{repoName, rev}] = resolver
+		return resolver
+	}
+
+	resolvers := make([]SearchResultResolver, 0, len(matches))
+	for _, match := range matches {
+		switch v := match.(type) {
+		case *result.FileMatch:
+			resolvers = append(resolvers, &FileMatchResolver{
+				db:           db,
+				FileMatch:    *v,
+				RepoResolver: getRepoResolver(v.Repo, ""),
+			})
+		case *result.RepoMatch:
+			resolvers = append(resolvers, getRepoResolver(v.RepoName(), v.Rev))
+		case *result.CommitMatch:
+			resolvers = append(resolvers, &CommitSearchResultResolver{
+				db:          db,
+				CommitMatch: *v,
+			})
+		}
+	}
+	return resolvers
 }
 
 func (sr *SearchResultsResolver) MatchCount() int32 {
